WorkSpace/requirements.txt
Flask
Flask-SQLALchemy
Flask-Login
Flask-Limiter
gunicorn
psycopg2-binary
python-dotenv
requests
itsdangerous
Werkzeug